SIMSBIOHUB-483: Working frontend pagination
