+++ api/src/paths/project/list.ts
@@ -160,9 +160,8 @@ export function getProjectList(): RequestHandler {
     defaultLog.debug({ label: 'getProjectList' });
 
     const page: number | undefined = req.query.page ? Number(req.query.page) : undefined;
-    const limit: number | undefined = req.query.limit ? Number(req.query.limit) : undefined;
+    const limit: number | undefined = req.query.limit ? Number(req.query.limit) : undefined
     const order: 'asc' | 'desc' | undefined = req.query.order ? (String(req.query.order) as 'asc' | 'desc') : undefined;
-
     const sort: string | undefined = req.query.sort ? String(req.query.sort) : undefined;
 
     const connection = getDBConnection(req['keycloak_token']);
