+++ app/src/interfaces/useObservationApi.interface.ts
@@ -1,16 +1,10 @@
 import { IObservationRecordWithSamplingData, ISupplementaryObservationData } from 'contexts/observationsTableContext';
+import { ApiPaginationResponseParams } from 'types/misc';
 
 export interface IGetSurveyObservationsResponse {
   surveyObservations: IObservationRecordWithSamplingData[];
   supplementaryObservationData: ISupplementaryObservationData;
-  pagination: {
-    total: number;
-    current_page: number;
-    last_page: number;
-    per_page?: number;
-    sort?: string;
-    order?: 'asc' | 'desc';
-  };
+  pagination: ApiPaginationResponseParams;
 }
 
 export interface IGetSurveyObservationsGeometryResponse {
