+++ app/src/interfaces/useProjectApi.interface.ts
@@ -3,6 +3,7 @@ import { PROJECT_PERMISSION, PROJECT_ROLE } from 'constants/roles';
 import { IProjectDetailsForm } from 'features/projects/components/ProjectDetailsForm';
 import { IProjectIUCNForm } from 'features/projects/components/ProjectIUCNForm';
 import { IProjectObjectivesForm } from 'features/projects/components/ProjectObjectivesForm';
+import { ApiPaginationResponseParams } from 'types/misc';
 
 export interface IGetProjectAttachment {
   id: number;
@@ -98,8 +99,8 @@ export interface IProjectSupplementaryData {
  * @interface IGetProjectsListResponse
  */
 export interface IGetProjectsListResponse {
-  projects: IProjectsListItemData[],
-  pagination: any; // TODO
+  projects: IProjectsListItemData[];
+  pagination: ApiPaginationResponseParams;
 }
 
 export interface IProjectsListItemData {
