+++ app/src/types/misc.ts
@@ -6,3 +6,12 @@ export type ApiPaginationOptions = {
   sort?: string;
   order?: 'asc' | 'desc';
 };
+
+export type ApiPaginationResponseParams = {
+  total: number;
+  current_page: number;
+  last_page: number;
+  per_page?: number;
+  sort?: string;
+  order?: 'asc' | 'desc';
+};
